+++ package.json
@@ -1,69 +1,40 @@
 {
-  "name": "project",
+  "name": "front-6th-chapter2-1",
   "private": true,
   "version": "0.0.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "start:basic": "vite serve --open ./index.basic.html",
+    "start:advanced": "vite serve --open ./index.advanced.html",
     "test": "vitest",
+    "test:basic": "vitest basic.test.js",
+    "test:advanced": "vitest advanced.test.js",
     "test:ui": "vitest --ui",
-    "test:coverage": "vitest run --coverage",
-    "build": "tsc -b && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "lint:fix": "eslint . --fix",
-    "prepare": "husky"
-  },
-  "dependencies": {
-    "@chakra-ui/icons": "^2.1.1",
-    "@chakra-ui/react": "^2.8.2",
-    "@dnd-kit/core": "^6.1.0",
-    "@dnd-kit/utilities": "^3.2.2",
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
-    "axios": "^1.7.3",
-    "framer-motion": "^11.3.2",
-    "msw": "^2.3.1",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "lint:basic": "pnpm eslint src/basic/**",
+    "lint:advanced": "pnpm eslint src/advanced/**"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^6.4.5",
-    "@testing-library/react": "^15.0.7",
-    "@testing-library/user-event": "^14.5.2",
-    "@types/express": "^4.17.21",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.13.1",
-    "@typescript-eslint/parser": "^7.13.1",
-    "@vitejs/plugin-react-swc": "^3.5.0",
-    "@vitest/coverage-v8": "^2.0.3",
-    "@vitest/ui": "^1.6.0",
-    "eslint-plugin-react-hooks": "^4.6.2",
-    "eslint-plugin-react-refresh": "^0.4.7",
-    "eslint": "^9.19.0",
-    "eslint-config-prettier": "^10.0.1",
-    "eslint-import-resolver-alias": "^1.1.2",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/user-event": "^14.6.1",
+    "@typescript-eslint/eslint-plugin": "^8.38.0",
+    "@typescript-eslint/parser": "^8.38.0",
+    "@vitest/ui": "^3.2.4",
+    "eslint": "^9.32.0",
+    "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-compat": "^6.0.2",
-    "eslint-plugin-cypress": "^2.13.3",
-    "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-prettier": "^5.2.3",
-    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-cypress": "^5.1.0",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-prettier": "^5.5.3",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-vitest": "^0.5.4",
-    "husky": "^9.1.7",
-    "globals": "^15.14.0",
-    "jsdom": "^25.0.1",
-    "lint-staged": "^15.4.3",
-    "prettier": "^3.4.2",
-    "vite-plugin-eslint": "^1.8.1",
-    "vitest": "^3.0.4",
-    "tsx": "^4.17.0",
-    "typescript": "^5.4.5",
-    "vite": "^5.3.1"
-  },
-  "lint-staged": {
-    "src/**/*.{js,jsx,ts,tsx}": [
-      "eslint --cache --fix",
-      "prettier --write"
-    ]
+    "globals": "^16.3.0",
+    "jsdom": "^26.1.0",
+    "prettier": "3.6.2",
+    "vite": "^7.0.5",
+    "vitest": "^3.2.4"
   }
 }
