+++ src/basic/utils/stores.js
@@ -1,0 +1,54 @@
+export const PRODUCTS = {
+  KEYBOARD: 'p1',
+  MOUSE: 'p2',
+  MONITOR_ARM: 'p3',
+  LAPTOP_POUCH: 'p4',
+  SPEAKER: 'p5',
+};
+export const productList = [
+  {
+    id: PRODUCTS.KEYBOARD,
+    name: '버그 없애는 키보드',
+    discountedPrice: 10000,
+    originalPrice: 10000,
+    quantity: 50,
+    onSale: false,
+    suggestSale: false,
+  },
+  {
+    id: PRODUCTS.MOUSE,
+    name: '생산성 폭발 마우스',
+    discountedPrice: 20000,
+    originalPrice: 20000,
+    quantity: 30,
+    onSale: false,
+    suggestSale: false,
+  },
+  {
+    id: PRODUCTS.MONITOR_ARM,
+    name: '거북목 탈출 모니터암',
+    discountedPrice: 30000,
+    originalPrice: 30000,
+    quantity: 20,
+    onSale: false,
+    suggestSale: false,
+  },
+  {
+    id: PRODUCTS.LAPTOP_POUCH,
+    name: '에러 방지 노트북 파우치',
+    discountedPrice: 15000,
+    originalPrice: 15000,
+    quantity: 0,
+    onSale: false,
+    suggestSale: false,
+  },
+  {
+    id: PRODUCTS.SPEAKER,
+    name: `코딩할 때 듣는 Lo-Fi 스피커`,
+    discountedPrice: 25000,
+    originalPrice: 25000,
+    quantity: 10,
+    onSale: false,
+    suggestSale: false,
+  },
+];
